Query board meta rather than lookup
The board metadata was captured in the metadata.boards table but it just
as fast to query most of that information on the fly instead of
duplicating that data into another table. This addresses that issue. The
only change not made was to the DB tables and trigger functions for now.