Support for postable_by for boards
This adds support for the new postable_by column on boards. This
includes updating the routes to allow setting this flag on create and
update, as well as returning this values on all read based methods. The
read based methods were also updated to not return boards that are
hidden when the hidePrivate option is set to true. A new DB method,
getBoardWriteAccess was created to handle returning the new flag on auth
checks while the getBoardInBoardMapping was refactored to handle boards
at a depth lower than 2 boards.